requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.15.2
filelock==3.16.1
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
openai==1.57.4
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
pypdfium2==4.30.0
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
sniffio==1.3.1
sqlparse==0.5.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
